more plot progress!
